+++ Assets/Asset/HQFPSTemplate/Core/Art/Meshes/_FP_Arms/Materials/Arm_Standard.mat
@@ -189,7 +189,7 @@ Material:
     - _DstBlend: 0
     - _EnableBlendModePreserveSpecularLighting: 1
     - _EnableFogOnTransparent: 1
-    - _FOV: 43
+    - _FOV: 44
     - _GlossMapScale: 0.5
     - _Glossiness: 0.5
     - _GlossyReflections: 1
