+++ Assets/UserFolder/Script/Scriptable/Scriptable Script/MeleeWeaponSoundScripatble.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cfeae6c39533ffd449937acad5bc07a9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
