+++ Assets/UserFolder/Script/Scriptable/Scriptable Script/RangeWeaponSoundScriptable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 53de1f3085ab9384c9a61ec6c6835814
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
