+++ Assets/UserFolder/Script/Scriptable/Scriptable Script/WeaponSoundScriptable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7104d6b027a1a748a8845389b526bc7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
