--- Assets/UserFolder/Script/Scriptable/Settings/Gun/AKM_Setting.asset
@@ -1,70 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 53de1f3085ab9384c9a61ec6c6835814, type: 3}
-  m_Name: AKM_Setting
-  m_EditorClassIdentifier: 
-  fireSound:
-  - {fileID: 8300000, guid: 3d32defdd09e2e044984e4d02afbe1a7, type: 3}
-  - {fileID: 8300000, guid: 3cc7d2fa1e3c8204cb90cf750dcc6fa0, type: 3}
-  emptySound:
-  - {fileID: 8300000, guid: 81fd9e79b0c4de8429890ca05996ac9f, type: 3}
-  changeModeSound:
-  - {fileID: 8300000, guid: 72971118ec16e97428f1ab90674b2e5f, type: 3}
-  equipSound:
-  - {fileID: 8300000, guid: 3013c91cb1efa464988c9407abaf9cd8, type: 3}
-  - {fileID: 8300000, guid: 0449fbe5d1232f8488e573b5e380189f, type: 3}
-  unequipSound:
-  - {fileID: 8300000, guid: 9271f804281c17847824a66eecfe71af, type: 3}
-  reloadSoundClips:
-  - audioClip: {fileID: 8300000, guid: 2b3281cd6a278a441a10e83e5b27e99b, type: 3}
-    delayTime: 0
-  - audioClip: {fileID: 8300000, guid: 5935f21e564c6984485383b5ff10402f, type: 3}
-    delayTime: 0.2
-  - audioClip: {fileID: 8300000, guid: 90f69d72569048f4ab10dc64d399c6e4, type: 3}
-    delayTime: 0.29
-  - audioClip: {fileID: 8300000, guid: 2a8e8423a1dd046499ddb162d7fc4d6e, type: 3}
-    delayTime: 0.04
-  - audioClip: {fileID: 8300000, guid: 2b3281cd6a278a441a10e83e5b27e99b, type: 3}
-    delayTime: 0.52
-  - audioClip: {fileID: 8300000, guid: 90f69d72569048f4ab10dc64d399c6e4, type: 3}
-    delayTime: 0.5
-  - audioClip: {fileID: 8300000, guid: 0fa0bd65d19bc174aac13ab86ee1e550, type: 3}
-    delayTime: 0.09
-  - audioClip: {fileID: 8300000, guid: e24b02f3f6c51ec469b9bef4d58c4847, type: 3}
-    delayTime: 0.035
-  - audioClip: {fileID: 8300000, guid: 9271f804281c17847824a66eecfe71af, type: 3}
-    delayTime: 0.425
-  emptyReloadSoundClips:
-  - audioClip: {fileID: 8300000, guid: 90f69d72569048f4ab10dc64d399c6e4, type: 3}
-    delayTime: 0.48
-  - audioClip: {fileID: 8300000, guid: 2a8e8423a1dd046499ddb162d7fc4d6e, type: 3}
-    delayTime: 0.04
-  - audioClip: {fileID: 8300000, guid: 90f69d72569048f4ab10dc64d399c6e4, type: 3}
-    delayTime: 0.98
-  - audioClip: {fileID: 8300000, guid: 0fa0bd65d19bc174aac13ab86ee1e550, type: 3}
-    delayTime: 0.1
-  - audioClip: {fileID: 8300000, guid: e24b02f3f6c51ec469b9bef4d58c4847, type: 3}
-    delayTime: 0.05
-  - audioClip: {fileID: 8300000, guid: 2b3281cd6a278a441a10e83e5b27e99b, type: 3}
-    delayTime: 0
-  - audioClip: {fileID: 8300000, guid: 9271f804281c17847824a66eecfe71af, type: 3}
-    delayTime: 0.35
-  - audioClip: {fileID: 8300000, guid: 19a94b9a688d358448157dcbcf469bd1, type: 3}
-    delayTime: 0.4
-  - audioClip: {fileID: 8300000, guid: b9578afcb3e05064397e31b1e9af7f99, type: 3}
-    delayTime: 0.1
-  - audioClip: {fileID: 8300000, guid: 69cc04f6384fb0f42a2f4101840a9480, type: 3}
-    delayTime: 0.05
-  - audioClip: {fileID: 8300000, guid: 2b3281cd6a278a441a10e83e5b27e99b, type: 3}
-    delayTime: 0.2
-  - audioClip: {fileID: 8300000, guid: 5935f21e564c6984485383b5ff10402f, type: 3}
-    delayTime: 0.15
