--- Assets/UserFolder/Script/Scriptable/Settings/Gun/AKM_Setting.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5115d5918e717c040aa7ad7d13f437a6
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
