+++ Assets/UserFolder/Script/Scriptable/Settings/Gun/AKM_SoundSetting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5115d5918e717c040aa7ad7d13f437a6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
