+++ Assets/UserFolder/Script/Scriptable/Settings/Gun/BaseballBat_SoundSetting.asset
@@ -1,0 +1,21 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 53de1f3085ab9384c9a61ec6c6835814, type: 3}
+  m_Name: BaseballBat_SoundSetting
+  m_EditorClassIdentifier: 
+  fireSound: []
+  emptySound: []
+  changeModeSound: []
+  equipSound: []
+  unequipSound: []
+  reloadSoundClips: []
+  emptyReloadSoundClips: []
