+++ Assets/UserFolder/Script/Scriptable/Settings/Gun/BaseballBat_SoundSetting.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6331ac65dc445d344b59c95076045682
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
