--- Assets/UserFolder/Script/Test/First Person Test/AnimationController_Test.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cbb4a9141278de6478532ddbe8f2c140
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
