--- Assets/UserFolder/Script/Test/First Person Test/AnimationGun_Test.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4484aa0df4dcf9742a2281c9d572027c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
