--- Assets/UserFolder/Script/Test/First Person Test/Animator Override/MeleeWeapon/FP_CombatKnife_Arm 1.overrideController
@@ -1,11 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!221 &22100000
-AnimatorOverrideController:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: FP_CombatKnife_Arm 1
-  m_Controller: {fileID: 0}
-  m_Clips: []
