--- Assets/UserFolder/Script/Test/First Person Test/Animator Override/MeleeWeapon/FP_CombatKnife_Arm 1.overrideController.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1d8805c693d2d3b408230bb7545a2bd1
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 22100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
