+++ Assets/UserFolder/Script/Test/First Person Test/Animator Override/MeleeWeapon/FP_CombatKnife_Arm.overrideController
@@ -7,5 +7,23 @@ AnimatorOverrideController:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_Name: FP_CombatKnife_Arm
-  m_Controller: {fileID: 0}
-  m_Clips: []
+  m_Controller: {fileID: 9100000, guid: 540bad8439832bc44bb4b8554834474c, type: 2}
+  m_Clips:
+  - m_OriginalClip: {fileID: 7400000, guid: 6fcd4ef4b0dd5df49a09e341d846c3d2, type: 2}
+    m_OverrideClip: {fileID: 7314049953401516684, guid: 4bc7819dab1a7a344bbf485737bb9ce1,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 1cee0a70c42f89f44ba4b62f9490d37d, type: 2}
+    m_OverrideClip: {fileID: 6757281245259898135, guid: 4bc7819dab1a7a344bbf485737bb9ce1,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 4840586a7b8754540ad3832bd81aa20f, type: 2}
+    m_OverrideClip: {fileID: 2745264705552615515, guid: 4bc7819dab1a7a344bbf485737bb9ce1,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 4966c1d6c082a3d46a96afb0b9716e91, type: 2}
+    m_OverrideClip: {fileID: 8753853386164614162, guid: 4bc7819dab1a7a344bbf485737bb9ce1,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: b367d6df9dfca974099bb40670f5ead9, type: 2}
+    m_OverrideClip: {fileID: -8628073155732118324, guid: 4bc7819dab1a7a344bbf485737bb9ce1,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 7192269aaafa4d54ea800b9cf70b323c, type: 2}
+    m_OverrideClip: {fileID: 7031575799557504332, guid: 4bc7819dab1a7a344bbf485737bb9ce1,
+      type: 3}
