+++ Assets/UserFolder/Script/Test/First Person Test/Animator Override/MeleeWeapon/FP_CombatKnife_Equipment.overrideController
@@ -1,0 +1,29 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!221 &22100000
+AnimatorOverrideController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: FP_CombatKnife_Equipment
+  m_Controller: {fileID: 9100000, guid: 540bad8439832bc44bb4b8554834474c, type: 2}
+  m_Clips:
+  - m_OriginalClip: {fileID: 7400000, guid: 6fcd4ef4b0dd5df49a09e341d846c3d2, type: 2}
+    m_OverrideClip: {fileID: 3580618316120492266, guid: 1a35791dcfb7ec9419a7149a5eb36145,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 1cee0a70c42f89f44ba4b62f9490d37d, type: 2}
+    m_OverrideClip: {fileID: 1876423795792452214, guid: 1a35791dcfb7ec9419a7149a5eb36145,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 4840586a7b8754540ad3832bd81aa20f, type: 2}
+    m_OverrideClip: {fileID: 4468177867845630117, guid: 1a35791dcfb7ec9419a7149a5eb36145,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 4966c1d6c082a3d46a96afb0b9716e91, type: 2}
+    m_OverrideClip: {fileID: 2062379842215227453, guid: 1a35791dcfb7ec9419a7149a5eb36145,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: b367d6df9dfca974099bb40670f5ead9, type: 2}
+    m_OverrideClip: {fileID: 6436263793849410671, guid: 1a35791dcfb7ec9419a7149a5eb36145,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 7192269aaafa4d54ea800b9cf70b323c, type: 2}
+    m_OverrideClip: {fileID: 2800935091575088496, guid: 1a35791dcfb7ec9419a7149a5eb36145,
+      type: 3}
