+++ Assets/UserFolder/Script/Test/First Person Test/Animator Override/MeleeWeapon/FP_CombatKnife_Equipment.overrideController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d54de72ef360f745a6f188f8f197583
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 22100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
