+++ Assets/UserFolder/Script/Test/First Person Test/Animator Override/MeleeWeapon/FP_Old Fire Axe_Arm.overrideController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7a3ade2717feef24182c0abcba5344b6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 22100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
