+++ Assets/UserFolder/Script/Test/First Person Test/Animator Override/MeleeWeapon/FP_Old Fire Axe_Equipment.overrideController
@@ -1,0 +1,26 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!221 &22100000
+AnimatorOverrideController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: FP_Old Fire Axe_Equipment
+  m_Controller: {fileID: 9100000, guid: 540bad8439832bc44bb4b8554834474c, type: 2}
+  m_Clips:
+  - m_OriginalClip: {fileID: 7400000, guid: 6fcd4ef4b0dd5df49a09e341d846c3d2, type: 2}
+    m_OverrideClip: {fileID: 3359366645728076354, guid: 6f4724007b10e004ea5c1da6ced1f294,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 1cee0a70c42f89f44ba4b62f9490d37d, type: 2}
+    m_OverrideClip: {fileID: -5730692241648192294, guid: 6f4724007b10e004ea5c1da6ced1f294,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 4840586a7b8754540ad3832bd81aa20f, type: 2}
+    m_OverrideClip: {fileID: -6359924051217592697, guid: 6f4724007b10e004ea5c1da6ced1f294,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: 4966c1d6c082a3d46a96afb0b9716e91, type: 2}
+    m_OverrideClip: {fileID: -1485273122184973416, guid: 6f4724007b10e004ea5c1da6ced1f294,
+      type: 3}
+  - m_OriginalClip: {fileID: 7400000, guid: b367d6df9dfca974099bb40670f5ead9, type: 2}
+    m_OverrideClip: {fileID: 1011378214786182432, guid: 6f4724007b10e004ea5c1da6ced1f294,
+      type: 3}
