+++ Assets/UserFolder/Script/Test/First Person Test/Animator Override/MeleeWeapon/FP_Old Fire Axe_Equipment.overrideController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ac2a4076c8054ef4b8fa83c13114f8be
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 22100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
