+++ Assets/UserFolder/Script/Test/First Person Test/LayerManager.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public static class LayerManager
+public class LayerManager
 {
-
+    
 }
