+++ Assets/UserFolder/Script/Test/First Person Test/Weapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ed028c5400c666d408783a8d0efa0e18
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
